Added Mailers Module

+++ app/views/users/partials/follow-form.blade.php
@@ -1,4 +1,4 @@
-
+@if ($signedIn)
     @if ($user->isFollowedBy($currentUser))
         {{ Form::open(['method' => 'DELETE', 'route' => ['follow_path', $user->id]]) }}
             {{ Form::hidden('userIdToUnfollow', $user->id) }}
@@ -10,3 +10,4 @@
             <button type="submit" class="btn btn-primary">Follow {{ $user->username }}</button>
         {{ Form::close() }}
     @endif
+@endif
